src/api_getway/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
httpx==0.28.1
python-jose==3.4.0
python-dotenv==1.1.0
slowapi==0.1.9
